Add files via upload
Prompts the user via a menu to create an instance of one of the shapes. Uses polymorphism to store shapes into a list. Continues to prompt until the user selects the exit option. Displays a neat, orderly table of all instantiated shapes, their dimensions, areas and, in the case of three-dimensional shapes, their volumes.