Add files via upload
The purpose of this assignment is to gain experience creating a database driven application using PHP server side code. The exercise is intended to demonstrate how even a small amount of database driven support can produce significant web site experiences for our users. You will be creating the server side support for an "infinite scroll" web page. On page load and at specific times when the user is scrolling on the page, an event will be triggered that causes an AJAX query to be sent to the server.